Refactor shell listing related code
